+++ src/main/scala/com/dataintuitive/viash/functionality/resources/JavaScriptScript.scala
@@ -27,9 +27,9 @@ case class JavaScriptScript(
   text: Option[String] = None,
   dest: Option[String] = None,
   is_executable: Option[Boolean] = Some(true),
-  parent: Option[URI] = None
+  parent: Option[URI] = None,
+  oType: String = "javascript_script"
 ) extends Script {
-  val `type` = "javascript_script"
   val meta = JavaScriptScript
   def copyResource(path: Option[String], text: Option[String], dest: Option[String], is_executable: Option[Boolean], parent: Option[URI]): Resource = {
     copy(path = path, text = text, dest = dest, is_executable = is_executable, parent = parent)
@@ -73,6 +73,7 @@ case class JavaScriptScript(
 object JavaScriptScript extends ScriptObject {
   val commentStr = "//"
   val extension = "js"
+  val oType = "javascript_script"
 
   def command(script: String): String = {
     "node \"" + script + "\""