+++ src/main/scala/com/dataintuitive/viash/platforms/NextFlowPlatform.scala
@@ -40,14 +40,10 @@ case class NextFlowPlatform(
   path: Option[String] = None,
   label: Option[String] = None,
   labels: List[String] = Nil,
-  stageInMode: Option[String] = None
+  stageInMode: Option[String] = None,
+  oType: String = "nextflow"
 ) extends Platform {
-  val `type` = "nextflow"
-  val hasSetup = false
-
   assert(version.isEmpty, "nextflow platform: attribute 'version' is deprecated")
-
-  val requirements: List[Requirements] = Nil
 
   private val nativePlatform = NativePlatform(id = id)
 
@@ -76,34 +72,35 @@ case class NextFlowPlatform(
     val allPars = functionality.arguments
 
     def inputFileExtO = allPars
-      .filter(_.`type` == "file")
+      .filter(_.isInstanceOf[FileObject])
       .find(_.direction == Input)
       .flatMap(_.default.map(_.toString.split('.').last))
 
     def inputs = allPars
-      .filter(_.`type` == "file")
+      .filter(_.isInstanceOf[FileObject])
       .count(_.direction == Input)
     def outputs = allPars
-      .filter(_.`type` == "file")
+      .filter(_.isInstanceOf[FileObject])
       .count(_.direction == Output)
 
     // All values for arguments/parameters are defined in the root of
     // the params structure. the function name is prefixed as a namespace
     // identifier. A "__" is used to separate namespace and arg/option.
+    //
+    // Required arguments also get a params.<argument> entry so that they can be
+    // called using --param value when using those standalone.
 
     val namespacedParameters: List[ConfigTuple] = {
       functionality.arguments.flatMap { dataObject => (dataObject.required, dataObject.default) match {
-        case (true, Some(x)) =>
+        case (true, _) =>
+          println(">>> Warning: " + dataObject.plainName + " is set to be required")
+          println(">>>          This will cause issues with NextFlow if this parameter is not provided explicitly.")
           Some(
             namespacedValueTuple(
               dataObject.plainName.replace("-", "_"),
-              x.toString
+              s"$${params.${dataObject.plainName}}"
             )(fun)
           )
-        case (true, None) =>
-          println(">>> Warning: " + dataObject.plainName + " is set to be required, but has no default value.")
-          println(">>>          This will cause issues with NextFlow if this parameter is not provided explicitly.")
-          None
         case (false, Some(x)) =>
           Some(
             namespacedValueTuple(
@@ -118,6 +115,16 @@ case class NextFlowPlatform(
               "no_default_value_configured"
             )(fun)
           )
+      }}
+    }
+
+    val requiredParameters: List[ConfigTuple] = {
+      functionality.arguments.flatMap { dataObject => (dataObject.required, dataObject.default) match {
+        case (true, _) =>
+          Some(
+            tupleToConfigTuple(dataObject.plainName.replace("-", "-") -> "viash_no_value")
+          )
+        case _ => None
       }}
     }
 
@@ -165,8 +172,9 @@ case class NextFlowPlatform(
     val asNestedTuples: List[ConfigTuple] = List(
       "docker.enabled" → true,
       "docker.runOptions" → "-i -v ${baseDir}:${baseDir}",
-      "process.container" → "dataintuitive/portash",
+      "process.container" → "dataintuitive/viash",
       "params" → NestedValue(
+        requiredParameters :::
         namespacedParameters :::
         List(
           tupleToConfigTuple("id" → ""),
@@ -196,7 +204,7 @@ case class NextFlowPlatform(
 
     val setup_main_outputFilters:String =
       allPars
-        .filter(_.`type` == "file")
+        .filter(_.isInstanceOf[FileObject])
         .filter(_.direction == Output)
         .map(par =>
           s"""
@@ -214,6 +222,21 @@ case class NextFlowPlatform(
             |
             |}""".stripMargin
           ).mkString("\n")
+
+    val setup_main_check =
+      s"""
+        |def checkParams(_params) {
+        |  _params.arguments.collect{
+        |    if (it.value == "viash_no_value") {
+        |      println("[ERROR] option --$${it.name} not specified in component ${fname}")
+        |      println("exiting now...")
+        |        exit 1
+        |    }
+        |  }
+        |}
+        |
+        |""".stripMargin
+
 
     val setup_main_utils =
       s"""
@@ -462,6 +485,8 @@ case class NextFlowPlatform(
         |
         |      def finalParams = overrideIO(newParams, inputs, outputs)
         |
+        |      checkParams(finalParams)
+        |
         |      new Tuple6(
         |        id,
         |        inputsForProcess,
@@ -504,7 +529,7 @@ case class NextFlowPlatform(
     val resultParseBlocks:List[String] =
       if (outputs >= 2) {
         allPars
-          .filter(_.`type` == "file")
+          .filter(_.isInstanceOf[FileObject])
           .filter(_.direction == Output)
           .map(par =>
             s"""
@@ -566,6 +591,7 @@ case class NextFlowPlatform(
     val setup_main = PlainFile(
       dest = Some("main.nf"),
       text = Some(setup_main_header +
+        setup_main_check +
         setup_main_utils +
         setup_main_outFromIn +
         setup_main_outputFilters +
@@ -655,11 +681,12 @@ object NextFlowUtils {
       valuePointer(dataObject.plainName.replace("-", "_"))
     }
 
+    // TODO: Should this not be converted from the json?
     quoteLong(dataObject.plainName) → NestedValue(
       tupleToConfigTuple("name" → dataObject.plainName) ::
       tupleToConfigTuple("otype" → dataObject.otype) ::
       tupleToConfigTuple("required" → dataObject.required) ::
-      tupleToConfigTuple("type" → dataObject.`type`) ::
+      tupleToConfigTuple("type" → dataObject.oType) ::
       tupleToConfigTuple("direction" → dataObject.direction.toString) ::
       tupleToConfigTuple("multiple" → dataObject.multiple) ::
       tupleToConfigTuple("multiple_sep" -> dataObject.multiple_sep) ::