+++ src/main/scala/com/dataintuitive/viash/platforms/Platform.scala
@@ -25,14 +25,15 @@ import requirements._
 import com.dataintuitive.viash.config.Version
 
 trait Platform {
-  val `type`: String
+  val oType: String
   val id: String
 
-  val hasSetup: Boolean
+  val hasSetup: Boolean = false
+  val requirements: List[Requirements] = Nil
 
   def modifyFunctionality(functionality: Functionality): Functionality
 
-  val requirements: List[Requirements]
+
 }
 
 object Platform {