+++ src/main/scala/com/dataintuitive/viash/platforms/requirements/NextFlowRequirements.scala
@@ -19,10 +19,9 @@ package com.dataintuitive.viash.platforms.requirements
 
 case class NextFlowRequirements(
   executor: Option[String] = None,
-  publishSubdir: Option[Boolean]
+  publishSubdir: Option[Boolean],
+  oType: String = "nextflow"
 ) extends Requirements {
-  val `type` = "nextflow"
-
   def installCommands: List[String] = {
     Nil
   }