+++ src/viash/viash_push/script.sh
@@ -1,12 +1,10 @@
 #!/bin/bash
 
 if [ "$par_mode" == "release" ]; then
-  echo "In release mode..."
-  if [ "$par_version" == "dev" ]; then
-    echo "For a release, you have to specify an explicit version using --version"
+  echo "In release mode with tag '$par_tag'."
+  if [ "$par_tag" == "dev" ]; then
+    echo "For a release, you have to specify an explicit version using --tag"
     exit 1
-  else
-    echo "Using version $par_version" to tag containers
   fi
 fi
 
@@ -38,6 +36,7 @@ if [[ $par_force == true ]]; then
     echo "No container push can and should be performed in this mode"
   elif [ "$par_mode" == "integration" ]; then
     "$par_viash" ns build \
+      -s "$par_src" \
       --platform "docker" \
       --query "$par_query" \
       --query_name "$par_query_name" \
@@ -50,14 +49,15 @@ if [[ $par_force == true ]]; then
       -c '.platforms[.type == "docker" || .type == "nextflow"].namespace_separator := "'$par_namespace_separator'"' \
       -c "$par_config_mod" \
       -l \
-      --setup --push | tee "$par_log"
+      --setup "push" | tee "$par_log"
   elif [ "$par_mode" == "release" ]; then
     "$par_viash" ns build \
+      -s "$par_src" \
       --platform "docker" \
       --query "$par_query" \
       --query_name "$par_query_name" \
       --query_namespace "$par_query_namespace" \
-      -c '.functionality.version := "'"$par_version"'"' \
+      -c '.functionality.version := "'"$par_tag"'"' \
       -c '.platforms[.type == "docker"].target_registry := "'"$par_registry"'"' \
       -c '.platforms[.type == "docker"].setup_strategy := "donothing"' \
       -c '.platforms[.type == "docker"].push_strategy := "alwayspush"' \
@@ -65,7 +65,7 @@ if [[ $par_force == true ]]; then
       -c '.platforms[.type == "docker" || .type == "nextflow"].namespace_separator := "'$par_namespace_separator'"' \
       -c "$par_config_mod" \
       -l \
-      --setup --push | tee "$par_log"
+      --setup "push" | tee "$par_log"
   else
     echo "Not a valid mode argument"
   fi
@@ -74,6 +74,7 @@ else
     echo "No container push can and should be performed in this mode"
   elif [ "$par_mode" == "integration" ]; then
     "$par_viash" ns build \
+      -s "$par_src" \
       --platform "docker" \
       --query "$par_query" \
       --query_name "$par_query_name" \
@@ -85,21 +86,22 @@ else
       -c '.platforms[.type == "docker" || .type == "nextflow"].namespace_separator := "'$par_namespace_separator'"' \
       -c "$par_config_mod" \
       -l \
-      --setup --push | tee "$par_log"
+      --setup "push" | tee "$par_log"
   elif [ "$par_mode" == "release" ]; then
     "$par_viash" ns build \
+      -s "$par_src" \
       --platform "docker" \
       --query "$par_query" \
       --query_name "$par_query_name" \
       --query_namespace "$par_query_namespace" \
-      -c '.functionality.version := "'"$par_version"'"' \
+      -c '.functionality.version := "'"$par_tag"'"' \
       -c '.platforms[.type == "docker"].target_registry := "'"$par_registry"'"' \
       -c '.platforms[.type == "docker"].setup_strategy := "donothing"' \
       -c '.platforms[.type == "nextflow"].registry := "'"$par_registry"'"' \
       -c '.platforms[.type == "docker" || .type == "nextflow"].namespace_separator := "'$par_namespace_separator'"' \
       -c "$par_config_mod" \
       -l \
-      --setup --push | tee "$par_log"
+      --setup "push" | tee "$par_log"
   else
     echo "Not a valid mode argument"
   fi